Update SQL scripts with required changes for the INSPIRE SOS: - Update create, drop, clear scripts - Create/Update update script from SOS 4.1.0 database model

+++ misc/db/Oracle/series/Oracle_series_update_40_to_41.sql
@@ -29,8 +29,8 @@
 
 ALTER TABLE series ADD firstTimeStamp timestamp;
 ALTER TABLE series ADD lastTimeStamp timestamp;
-ALTER TABLE series ADD firstNumericValue number(19,2);
-ALTER TABLE series ADD lastNumericValue number(19,2);
+ALTER TABLE series ADD firstNumericValue DOUBLE PRECISION;
+ALTER TABLE series ADD lastNumericValue DOUBLE PRECISION;
 ALTER TABLE series ADD unitId number(19,0);
 
 alter table series add constraint seriesUnitFk foreign key (unitId) references unit;
